refactor: split app cli api module
